Implement listen matcher based on database
It still doesn't do anything with the matches, but I could make a new
table to store them. For now this is still for experimenting, and how to
get good matches. Currently it matches 89% of my listens that are not
yet in the database, which is quite good considering that many of the
tracks are not even in my library. From a glance, the next source of
misses are when the title contains a feat. artist or something like
"[Deluxe Edition]" or "[Bonus Track]", those should be easy to strip.