Add some ideas for batter matching based on misses
Sometimes the title differs by "A" or "The" being present or not. In one
case, the listen is spelled with "and" but the library entry with "&".
Regularly one of the two is spelled with "Pt." while the other contains
"Part" written out. In some cases there is a difference in punctuation
("Panic! at the Disco" vs. "Panic at the Disco"). All of this could be
matched relatively easily if we just delete the words that convey little
information and normalize some others.