Replace Guetzli with cjpegli for thumbnail compression
To target the same file size as before, I deleted all thumbnails from
my database for file ids that are 0 or 1 modulo 10. This comprises 369
covers. To be sure I recomputed the Guetzli covers because I had tweaked
the resizing args in the past. Then I did a binary search for the
Butteraugli distance that produces the same size thumbnails. We can
check the size with the following query:

    select
      sum(length(data)),
      sum(length(data)) / count(*)
    from
      thumbnails
    where
      file_id % 10 < 2;

The results are as follows:

compressor  quality  cum. bytes  avg. bytes  size ratio  secs.
----------  -------  ----------  ----------  ----------  ------
guetzli     q=97      4,267,771      11,565      100.0%  166.88
cjpegli     d=1.000   2,782,842       7,541       65.2%  102.78
cjpegli     d=0.500   4,059,082      11,000       95.1%  103.70
cjpegli     d=0.470   4,195,241      11,369       98.3%  102.51
cjpegli     d=0.459   4,228,262      11,458       99.1%  102.60  <--
cjpegli     d=0.250   5,601,187      15,179      131.2%  105.60

The cumulative bytes is the combined size of all 369 thumbnails while
the average bytes is the average size of one thumbnail. The initial
quality value for guetzli was chosen in a comparision with other
compressors but I don't remember having a good reason for a particular
absolute quality level aside from just using 95-100 quality level to not
get too many artifacts, so probably d=1.0 would have been fine, but we
target the original size, so with d=0.459 we get pretty close. I also
timed how long it took, but only a single measurement per size so take
them with a grain of salt. (This is with 16 threads.) Still, cjpegli
is about a minute faster than guetzli was so that is nice, and the
difference will be even more pronounced on a device like a Raspberry Pi.