Switch thumbnail compressor back to Guetzli
This is very hard to tell and somewhat subjective, but it seems Guetzli
produces better images after all. Though maybe I need to disable the
progressive scan? It doesn't matter so much for thumbnails, we load them
before they scroll into view and there is a fallback image. Let's give
that a final try then.