Begin working on using only the Alsa HW devices
I'm struggling with this noise issue on my UMC404HD, and I think it's a
bug in Alsa. I wonder if sidestepping the "plug:" device fixes this, let
me get as close to the raw underlying device as I can get.

While at it, I can move the high-pass filter into the playback thread.
I'll do that in a follow-up commit.