Restructure how the filter and decoder interact
Previously we would apply the filter right at the decoding stage. I
think my reasoning was that I wanted to do as little computation as
possible in the thread that fills the Alsa buffer (and it holds the
player lock while doing so). However, moving the filter there has some
advantages:

 * We can choose the filter cutoff very late, just before we send out
   the audio data, so we can make it user-controllable at runtime, just
   like the volume.
 * We can choose to always output in 16 bits, regardless of the input
   bit depth, since we go through f32 anyway. This simplifies
   configuring the audio device, we have to switch the format less
   often, now we only need to switch it when the sample rate changes.

It does have the downside that we hold the lock for longer. But now that
I'm revisiting this code, we can avoid a few memcopies, and Musium uses
very little CPU anyway when it's not decoding, so I think it should be
fine.

In this commit we do not yet introduce the state to change the cutoff
frequency at runtime, so there are some unused variable warnings now.