Pass audio device name directly to Alsa
Now that I wanted to use the hardware directly, it turns out the Realtek
in my laptop (and also the one in the Dock) only support 48 kHz, so I
still need the "plug" device to test on my laptop! But on my target
Raspberry I have the Behringer UMC404 connected, and it supports all
sample rates I need, so that one I could interface with directly.