Assert that player has no empty blocks
I hit this case again where playback hangs at the end of a track, that
I still haven't been able to debug after years, but now diving in with
LLDB, it looks like the peek_mut call that peeks the block, returns a
block where the data is empty. Then we write 0 samples, and we also
don't consume and move on, so the hang makes sense, but where do empty
blocks come from? That makes less sense, but probably a bug in the
decoder. Let's start with adding some assertions.