Do not enqueue empty blocks from the decoder
I have good hope that I now finally found the cause of the hang. Maybe
the decoder creates an empty block, when it happened to exactly decode
everything in the previous run, but it doesn't know that until we try
to decode more, and then we get an empty block.

Could it be that this finally fixes the bug?