Begin working on thumbnail color analysis
I noticed in search, in the discographies where it shows many images,
they take a few milliseconds to show up, and it makes them flicker into
view which is ugly. I already added a grey backdrop, but it's still
visible. I wonder if it would help to select a better background color
that is more aligned with the thumbnail. Then I got a bit nerd-sniped
on extracting that color, and I think I have a good algorithm now. Now
we need to run that and store it in the database, and then later we can
serve it in the webinterface in the response json.

I suspect the reason for the flicker is not loading the image, because
they are cached, but either decoding or downscaling it. I suppose the
browser evicts it from memory and then has to decode it again.

This commit adds an analysis step to the thumbnail pipeline, but it does
not do anything with that information yet.