Add the notion of a time vector
I have this idea in my head for months, let's actually try it this time.
Music is very seasonal for me, I have tracks that are better suited for
a hot summer night, tracks that are better suited for a grey autumn day.
Also, time of the week matters, I listen to chill music on a Saturday
or Sunday morning, I listen to bangers more in the evenings. So I
thought, what if we create a moving average for a track (and album, and
artist) to track the time of the year, week, and day? Then we can make
better suggestions for what to listen to based on the current time.

This adds the types for that, and also a mapping of weekdays to the
circle, based on me toying around with Geogebra.