Serve time-adjusted discover score from API
This integrates the new "for now"-weighed discover scores, and it also
exposes a more raw "for now" score that needs to be integrated into the
UI still.

This includes some rustfmt reformattings, but not everything it would
reformat in the same files; some changes are very invasive and I want
to avoid doing that here.