Optimize the reload stage after a scan
Still not as efficient as it could be, but at least we avoid rebuilding
the entire index, and this is probably fast enough.

Benchmark results when running `musium scan` on an up-to-date database,
which does some work to list files (but they are cached in the page
cache, the first run takes multiple seconds as my disk rattles, but
subsequent runs are very fast), but spends most of its time building
the index from the database:

Benchmark 1: /tmp/before scan musium.conf
  Time (mean ± σ):     651.1 ms ±   3.0 ms    [User: 573.8 ms, System: 118.8 ms]
  Range (min … max):   645.7 ms … 657.1 ms    20 runs

Benchmark 2: /tmp/after scan musium.conf
  Time (mean ± σ):     385.7 ms ±   1.9 ms    [User: 320.5 ms, System: 107.0 ms]
  Range (min … max):   382.1 ms … 389.6 ms    20 runs

Summary
  /tmp/after scan musium.conf ran
    1.69 ± 0.01 times faster than /tmp/before scan musium.conf

So this is a worthwhile change! Still, it shows that building the index
is pretty fast, under 400ms on my laptop for 20k tracks. It'll be a lot
slower on the Raspberry Pi, but still, no need to optimize it a lot
further at this stage.